Update actuatorLineSource.C
It is a modification made on the basis of the library named thesis-FloatingTurbine, added function for active pitch of floating wind turbines. Line 544 introduces a new function, windPitching, which is a modified version of the original harmonicPitching() function. At each time step, it calculates the relative wind speed at the hub height based on the free-stream velocity, the surge velocity of the floating platform, and its pitch angular velocity, and adjusts the blade pitch angle accordingly. But the specific relative wind speed threshold for variable pitch is directly written in the code and only applies to the IEA 15 MW floating wind turbine.